Implement post logic
* add post methods

+++ app/persistence/src/main/java/com/fisherman/companion/persistence/PostRepositoryJdbcImpl.java
@@ -6,12 +6,16 @@ import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
+import com.fisherman.companion.dto.BoundingBoxDimensions;
 import com.fisherman.companion.dto.CategoryDto;
 import com.fisherman.companion.dto.PostDto;
 import com.fisherman.companion.dto.PostStatus;
@@ -24,33 +28,35 @@ public class PostRepositoryJdbcImpl implements PostRepository {
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
     @Override
-    public void savePost(final PostDto postDto) {
+    public Long savePost(final PostDto postDto) {
         final String sql = """
                 INSERT INTO posts (user_id, category_id, title, description, start_date, latitude, longitude, contact_info, status)
                 VALUES (:userId, :categoryId, :title, :description, :startDate, :latitude, :longitude, :contactInfo, :status)
                 """;
 
         final MapSqlParameterSource params = new MapSqlParameterSource()
-                .addValue("userId", postDto.userId())
-                .addValue("categoryId", postDto.category().id())
-                .addValue("title", postDto.title())
-                .addValue("description", postDto.description())
-                .addValue("startDate", postDto.startDate())
-                .addValue("latitude", postDto.latitude())
-                .addValue("longitude", postDto.longitude())
-                .addValue("contactInfo", postDto.contactInfo())
-                .addValue("status", postDto.status().name());
+                .addValue("userId", postDto.getUserId())
+                .addValue("categoryId", postDto.getCategory().getId())
+                .addValue("title", postDto.getTitle())
+                .addValue("description", postDto.getDescription())
+                .addValue("startDate", postDto.getStartDate())
+                .addValue("latitude", postDto.getLatitude())
+                .addValue("longitude", postDto.getLongitude())
+                .addValue("contactInfo", postDto.getContactInfo())
+                .addValue("status", postDto.getStatus().name());
 
-        namedParameterJdbcTemplate.update(sql, params);
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+
+        namedParameterJdbcTemplate.update(sql, params, keyHolder);
+
+        return Objects.requireNonNull(keyHolder.getKey()).longValue();
     }
 
     @Override
     public List<PostDto> findAllCategoriesPosts(final int take, final int skip, final LocalDateTime timeToFilter) {
         final String sql = """
-                SELECT p.id, p.user_id, p.category_id, p.title, p.description, p.start_date, p.latitude, p.longitude, p.contact_info, p.status, c.name as category_name
-                FROM posts p
-                LEFT JOIN categories c ON p.category_id = c.id
-                WHERE p.status = 'opened' AND p.start_date >= :startDate
+                SELECT * FROM posts p
+                WHERE p.status = 'open' AND p.start_date >= :startDate
                 ORDER BY p.start_date
                 LIMIT :take
                 OFFSET :skip
@@ -66,6 +72,49 @@ public class PostRepositoryJdbcImpl implements PostRepository {
     }
 
     @Override
+    public List<PostDto> findPostsInBoundingBoxByCategory(final BoundingBoxDimensions boxDimensions, final LocalDateTime from, final LocalDateTime to, final Long categoryId) {
+        String sql = """
+                SELECT * FROM posts
+                WHERE status = 'open'
+                    AND start_date BETWEEN :from AND :to
+                    AND latitude BETWEEN :minLat AND :maxLat
+                    AND longitude BETWEEN :minLng AND :maxLng
+                    AND category_id = :category
+                ORDER BY start_date;
+                """;
+        final Map<String, Object> params = Map.of(
+                "minLat", boxDimensions.minLat(),
+                "minLng", boxDimensions.minLng(),
+                "maxLat", boxDimensions.maxLat(),
+                "maxLng", boxDimensions.maxLng(),
+                "from", Timestamp.valueOf(from),
+                "to", Timestamp.valueOf(to),
+                "category", categoryId
+        );
+
+        return namedParameterJdbcTemplate.query(sql, params, new PostMapper());
+    }
+
+    @Override
+    public List<PostDto> findUserPosts(final Long userId, final int take, final int skip) {
+        final String sql = """
+                SELECT *
+                FROM posts
+                WHERE user_id = :userId
+                ORDER BY start_date DESC
+                LIMIT :take
+                OFFSET :skip
+                """;
+
+        final MapSqlParameterSource params = new MapSqlParameterSource()
+                .addValue("userId", userId)
+                .addValue("take", take)
+                .addValue("skip", skip);
+
+        return namedParameterJdbcTemplate.query(sql, params, new PostMapper());
+    }
+
+    @Override
     public void updatePostById(final PostDto postDto) {
         final String sql = """
                 UPDATE posts
@@ -76,50 +125,58 @@ public class PostRepositoryJdbcImpl implements PostRepository {
                     start_date = COALESCE(:startDate, start_date),
                     latitude = COALESCE(:latitude, latitude),
                     longitude = COALESCE(:longitude, longitude),
-                    contact_info = COALESCE(:contactInfo, contact_info),
-                    status = COALESCE(:status, status)
-                WHERE id = :id
+                    contact_info = COALESCE(:contactInfo, contact_info)
+                WHERE id = :id AND status = 'open'
                 """;
 
         final MapSqlParameterSource params = new MapSqlParameterSource()
-                .addValue("postId", postDto.id())
-                .addValue("categoryId", postDto.category().id())
-                .addValue("title", postDto.title())
-                .addValue("description", postDto.description())
-                .addValue("startDate", postDto.startDate())
-                .addValue("latitude", postDto.latitude())
-                .addValue("longitude", postDto.longitude())
-                .addValue("contactInfo", postDto.contactInfo())
-                .addValue("status", postDto.status().name())
-                .addValue("id", postDto.id());
+                .addValue("userId", postDto.getUserId())
+                .addValue("categoryId", postDto.getCategory().getId())
+                .addValue("title", postDto.getTitle())
+                .addValue("description", postDto.getDescription())
+                .addValue("startDate", postDto.getStartDate())
+                .addValue("latitude", postDto.getLatitude())
+                .addValue("longitude", postDto.getLongitude())
+                .addValue("contactInfo", postDto.getContactInfo())
+                .addValue("id", postDto.getId());
 
         namedParameterJdbcTemplate.update(sql, params);
     }
 
     @Override
-    public void deleteById(final Long id) {
+    public void deleteById(final Long id, final Long userId) {
         final String sql = """
-                DELETE FROM posts WHERE id = :id
+                DELETE FROM posts WHERE id = :id and user_id = :userId
                 """;
 
-        namedParameterJdbcTemplate.update(sql, Map.of("id", id));
+        Map<String, Object> params = Map.of(
+                "id", id,
+                "userId", userId
+        );
+        namedParameterJdbcTemplate.update(sql, params);
     }
 
     private static class PostMapper implements RowMapper<PostDto> {
         @Override
         public PostDto mapRow(final ResultSet rs, final int rowNum) throws SQLException {
-            return PostDto.builder()
-                          .id(rs.getLong("id"))
-                          .userId(rs.getLong("user_id"))
-                          .category(new CategoryDto(rs.getLong("category_id"), rs.getString("category_name")))
-                          .title(rs.getString("title"))
-                          .description(rs.getString("description"))
-                          .startDate(rs.getTimestamp("start_date").toLocalDateTime())
-                          .latitude(rs.getDouble("latitude"))
-                          .longitude(rs.getDouble("longitude"))
-                          .contactInfo(rs.getString("contact_info"))
-                          .status(PostStatus.valueOf(rs.getString("status")))
-                          .build();
+            final CategoryDto category = new CategoryDto();
+
+            category.setId(rs.getLong("category_id"));
+
+            final PostDto post = new PostDto();
+
+            post.setId(rs.getLong("id"));
+            post.setUserId(rs.getLong("user_id"));
+            post.setCategory(category);
+            post.setTitle(rs.getString("title"));
+            post.setDescription(rs.getString("description"));
+            post.setStartDate(rs.getTimestamp("start_date").toLocalDateTime());
+            post.setLatitude(rs.getDouble("latitude"));
+            post.setLongitude(rs.getDouble("longitude"));
+            post.setContactInfo(rs.getString("contact_info"));
+            post.setStatus(PostStatus.valueOf(rs.getString("status").toUpperCase()));
+
+            return post;
         }
     }
 }
